+move module to directory and split off tests, fix tty container name
